Fix -Wsign-conversion warnings in Board, Engine, search
- Board.cpp:759  history[static_cast<size_t>(i)] — loop index is int
- Engine.cpp:54  tt.resize(static_cast<size_t>(desired_hash_mb))
- search.cpp:599 move_list[static_cast<size_t>(i)].score — loop index is int

All three were safe at runtime (values always non-negative) but mixing
signed loop counters with unsigned container subscripts is technically
undefined behaviour and correctly flagged by -Wsign-conversion.

+++ src/Board.cpp
@@ -756,7 +756,7 @@ bool Board::is_draw() const {
     int reps = 0;
     int stop = std::max(0, (int)history.size() - halfmove_clock);
     for (int i = (int)history.size() - 2; i >= stop; i -= 2) {
-        if (history[i].hash == hash) {
+        if (history[static_cast<size_t>(i)].hash == hash) {
             reps++;
             if (reps >= 1) return true; // 2-fold (current + one prior)
         }
